// Pparty/Implemento-Last/src/app/model/festa.ts
export class Festa {
    idFesta: number;
    data: string;
    titulo: string;
    enderecofesta: string;
    valorIngresso: number;
    cardapio: string;
    qingresso: number;
    categoria: string;
    latitude: number;
    longitude: number;
    idUsuariof: number;

    constructor()
    {
        this.idFesta = 0;
        this.data = "";
        this.titulo = "";
        this.enderecofesta = "";
        this.valorIngresso = 0;
        this.cardapio = "";
        this.qingresso = 0;
        this.categoria = "";
        this.latitude = 0;
        this.longitude = 0;
        this.idUsuariof = 0;
    }
}